lib: split common to common and internal packages
