+++ lib/backend/repository/repository.go
@@ -26,6 +26,7 @@ import (
 	"path"
 
 	"github.com/appc/docker2aci/lib/common"
+	"github.com/appc/docker2aci/lib/internal"
 	"github.com/appc/docker2aci/lib/types"
 	"github.com/appc/spec/schema"
 )
@@ -60,7 +61,7 @@ func NewRepositoryBackend(username string, password string, insecure bool) *Repo
 }
 
 func (rb *RepositoryBackend) GetImageInfo(url string) ([]string, *types.ParsedDockerURL, error) {
-	dockerURL, err := common.ParseDockerURL(url)
+	dockerURL, err := internal.ParseDockerURL(url)
 	if err != nil {
 		return nil, nil, err
 	}
